various d_bg debug cleanup / d_menu_window_HIO debug (#3132)
* various d_bg debug cleanup

* matching forward decl for compiler compatibility

* fix build

+++ include/SSystem/SComponent/c_bg_w.h
@@ -18,7 +18,7 @@ public:
     cBgW_BgId() { Ct(); }
     void Ct() { m_id = 0x100; }
 
-    u16 GetId() const { return m_id; }
+    int GetId() const { return m_id; }
 };
 
 bool cBgW_CheckBGround(float a1);
